+++ CHANGES.md
@@ -1,9 +1,10 @@
 Release notes
 =============
 
-2025081400
+2025081401
 
 1. 2025-08-14: fix: #336: deleted users are excluded from everywhere when searching by and merging users.
+2. 2025-08-14: improvement: #281: add debug info into log when no course module is found when regrading. Thanks to @Richardvi.
 
 2025081300